line/requirements.txt
opencv-python
PyYAML
shapely==1.8.5.post1
monai==0.7.0
tqdm
scipy
imageio
pyvista
geopandas
ignite==1.1.0
pytorch-ignite==0.4.9